+++ Source/Examples/Xamarin.Forms/IssueDemos/IssueDemos/App.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+using Xamarin.Forms;
+
+namespace IssueDemos
+{
+    public class App : Application
+    {
+        public App()
+        {
+            MainPage = new NavigationPage(new MainPage());
+        }
+    }
+}